Refactored some things

+++ src/Controller/SnippetController.php
@@ -6,7 +6,6 @@ use App\Entity\Snippet;
 use App\Form\SnippetFormType;
 use App\Repository\SnippetRepository;
 use App\Service\SnippetHelper;
-use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,15 +14,15 @@ use Symfony\Component\Routing\Annotation\Route;
 class SnippetController extends AbstractController
 {
     /**
-     * @Route({"de": "/textbausteine", "en": "/snippets"}, name="snippets")
+     * @Route({"de": "/textbausteine", "en": "/snippets"}, name="snippet_index")
      * @IsGranted("ROLE_USER")
      */
-    public function listSnippets(Request $request, SnippetRepository $snippetRepository, SnippetHelper $snippetHelper)
+    public function index(Request $request, SnippetRepository $snippetRepository, SnippetHelper $snippetHelper)
     {
         $snippets = $snippetRepository->findOrderedByPriority();
         $snippets = $snippetHelper->sortSnippets($snippets);
 
-        return $this->render('snippets/list.html.twig', [
+        return $this->render('snippets/index.html.twig', [
             'snippets' => $snippets,
         ]);
     }
@@ -32,7 +31,7 @@ class SnippetController extends AbstractController
      * @Route({"de": "/textbausteine/neu", "en": "/snippets/new"}, name="snippet_new")
      * @IsGranted("ROLE_USER")
      */
-    public function newSnippet(Request $request, EntityManagerInterface $em)
+    public function new(Request $request)
     {
         $form = $this->createForm(SnippetFormType::class);
         $form->handleRequest($request);
@@ -40,12 +39,13 @@ class SnippetController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $snippet = $form->getData();
 
+            $em = $this->getDoctrine()->getManager();
             $em->persist($snippet);
             $em->flush();
 
             $this->addFlash('success', 'msg.snippet.added');
 
-            return $this->redirectToRoute('snippets');
+            return $this->redirectToRoute('snippet_index');
         }
 
         return $this->render('snippets/new.html.twig', [
@@ -58,18 +58,19 @@ class SnippetController extends AbstractController
      * @Route({"de": "/textbausteine/{id}", "en": "/snippets/{id}"}, name="snippet_edit")
      * @IsGranted("ROLE_USER")
      */
-    public function editSnippet(Snippet $snippet, Request $request, EntityManagerInterface $em)
+    public function edit(Snippet $snippet, Request $request)
     {
         $form = $this->createForm(SnippetFormType::class, $snippet);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
             $em->persist($snippet);
             $em->flush();
 
             $this->addFlash('primary', 'msg.snippet.edited');
 
-            return $this->redirectToRoute('snippets');
+            return $this->redirectToRoute('snippet_index');
         }
 
         return $this->render('snippets/edit.html.twig', [
@@ -82,18 +83,16 @@ class SnippetController extends AbstractController
      * @Route({"de": "/textbausteine/{id}/delete", "en": "/snippets/{id}/delete"}, name="snippet_delete")
      * @IsGranted("ROLE_USER")
      */
-    public function deleteSnippet(Snippet $snippet, Request $request, EntityManagerInterface $em)
+    public function delete(Snippet $snippet, Request $request)
     {
-        $em->remove($snippet);
-        $em->flush();
+        if ($this->isCsrfTokenValid('delete'.$snippet->getId(), $request->request->get('_token'))) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($snippet);
+            $em->flush();
+        }
 
         $this->addFlash('danger', 'msg.snippet.deleted');
 
-        return $this->redirectToRoute('snippets');
-    }
-
-    private function compareSnippets($obj1, $obj2)
-    {
-        return $obj1->getPriority() > $obj2->getPriority();
+        return $this->redirectToRoute('snippet_index');
     }
 }
